SegForestNet/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
albumentations>=1.3.0
matplotlib>=3.7.0
numpy>=1.23.5
scikit-learn>=1.2.1
tqdm>=4.65.0
pillow>=9.5.0
rasterio>=1.3.0
scikit-image>=0.19.0
opencv-python>=4.7.0 